EDU-1027 Fetch timestamp containing data on client side
