+++ src/components/pages/user.js
@@ -1,20 +1,31 @@
 import PropTypes from 'prop-types';
 import Markdown from '../markdown.js';
-import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import FavoriteStar from '../favorite-star.js';
 import DocumentCard from '../document-card.js';
 import ProfileHeader from '../profile-header.js';
+import React, { useEffect, useState } from 'react';
 import { FAVORITE_TYPE } from '../../domain/constants.js';
-import { documentMetadataShape, publicUserShape } from '../../ui/default-prop-types.js';
+import UserApiClient from '../../api-clients/user-api-client.js';
+import { publicUserShape } from '../../ui/default-prop-types.js';
+import { useSessionAwareApiClient } from '../../ui/api-helper.js';
 
 const CARD_BATCH_SIZE = 8;
 
 export default function User({ PageTemplate, initialState }) {
   const { t } = useTranslation('user');
-  const { user, documents } = initialState;
+  const userApiClient = useSessionAwareApiClient(UserApiClient);
 
+  const { user } = initialState;
+  const [documents, setDocuments] = useState([]);
   const [visibleDocumentsCount, setVisibleDocumentsCount] = useState(CARD_BATCH_SIZE);
+
+  useEffect(() => {
+    (async () => {
+      const userApiClientResponse = await userApiClient.getCreatedDocuments({ userId: user._id });
+      setDocuments(userApiClientResponse.documents);
+    })();
+  }, [user, userApiClient]);
 
   const handleMoreDocumentsClick = () => {
     setVisibleDocumentsCount(visibleDocumentsCount + CARD_BATCH_SIZE);
@@ -77,7 +88,6 @@ export default function User({ PageTemplate, initialState }) {
 User.propTypes = {
   PageTemplate: PropTypes.func.isRequired,
   initialState: PropTypes.shape({
-    user: publicUserShape.isRequired,
-    documents: PropTypes.arrayOf(documentMetadataShape).isRequired
+    user: publicUserShape.isRequired
   }).isRequired
 };
